requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-oauth-toolkit==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jwcrypto==1.5.6
kombu==5.4.1
msgpack==1.1.0
msgpack-python==0.5.6
oauthlib==3.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==75.0.0
six==1.16.0
sniffio==1.3.1
social-auth-core==4.5.4
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
zope.interface==7.0.3
